requirements.txt
pandas>=2.2
numpy>=1.26
pyyaml>=6.0
pyarrow>=15.0
